requirements.txt
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
apispec==6.6.1
apispec-webframeworks==1.1.0
arrow==1.3.0
asyncio==3.4.3
attrs==24.2.0
Automat==22.10.0
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.113
botocore==1.34.162
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cmdstanpy==1.2.4
colorama==0.4.6
constantly==23.10.4
contourpy==1.2.1
convertdate==2.4.0
cookiecutter==2.6.0
cycler==0.12.1
Cython==3.0.11
deepdiff==7.0.1
dnspython==2.6.1
ephem==4.1.5
et-xmlfile==1.1.0
fastapi==0.110.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
fonttools==4.53.1
frozendict==2.4.4
gevent==24.2.1
gevent-websocket==0.10.1
gitignore_parser==0.1.11
greenlet==3.0.3
h11==0.14.0
hijri-converter==2.3.1
holidays==0.54
html5lib==1.1
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.6
importlib_resources==6.4.3
incremental==24.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
kthread==0.2.3
LunarCalendar==0.0.9
lxml==5.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
mdurl==0.1.2
multitasking==0.0.11
networkx==3.3
numpy==1.26.4
openpyxl==3.1.2
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
passlib==1.7.4
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prophet==1.1.5
protobuf==5.28.0rc2
pyarrow==16.1.0
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
PyMeeus==0.5.12
pymongo==4.7.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.3
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
rich==13.7.1
s3transfer==0.10.2
scikit-learn==1.5.1
scipy==1.14.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.30
stanio==0.5.1
starlette==0.36.3
taipy==4.0.0.dev0
taipy-config==4.0.0.dev0
taipy-core==4.0.0.dev0
taipy-gui==4.0.0.dev0
taipy-rest==4.0.0.dev0
taipy-templates==4.0.0.dev0
tenacity==9.0.0
text-unidecode==1.3
threadpoolctl==3.5.0
toml==0.10.2
tqdm==4.66.5
Twisted==24.3.0
twisted-iocpsupport==1.0.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
upstox-python-sdk==2.5.0
urllib3==2.2.2
uuid==1.30
uvicorn==0.29.0
watchdog==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yfinance==0.2.41
zope.event==5.0
zope.interface==7.0.1
